fix(scripting): port Lua scripting to SDL3 and current main
Adjustments needed to bring the cherry-picked LuaJIT branch (SDL2-era,
Dec 2025) up to current main after the SDL3/GPU/TTF/panels waves:

- SDL2/SDL.h includes become SDL3/SDL.h (lua_core.c, lua_api.c)
- SDL_Keycode is unsigned in SDL3: lua_scripting_keydown/keyup now take
  uint32_t, matching amod_keydown/keyup
- SDL_SetClipboardText returns bool in SDL3 (was int 0 on success)
- INVENTORYSIZE/CONTAINERSIZE compile-time constants were replaced on
  main by runtime _inventorysize/_containersize (v3 vs v3.5 servers):
  bounds checks and the C.INVENTORYSIZE/C.CONTAINERSIZE Lua constants
  now use the runtime values (registered after server version selection)

+++ src/scripting/lua_core.c
@@ -16,7 +16,7 @@
 #include <lualib.h>
 #include <lauxlib.h>
 
-#include <SDL2/SDL.h>
+#include <SDL3/SDL.h>
 
 #include "astonia.h"
 #include "scripting/lua_interface.h"
@@ -820,14 +820,14 @@ int lua_scripting_mouse_click(int x, int y, int what)
 	return call_lua_handler_int("on_mouse_click", 3, x, y, what);
 }
 
-int lua_scripting_keydown(int key)
+int lua_scripting_keydown(uint32_t key)
 {
-	return call_lua_handler_int("on_keydown", 1, key);
+	return call_lua_handler_int("on_keydown", 1, (int)key);
 }
 
-int lua_scripting_keyup(int key)
+int lua_scripting_keyup(uint32_t key)
 {
-	return call_lua_handler_int("on_keyup", 1, key);
+	return call_lua_handler_int("on_keyup", 1, (int)key);
 }
 
 int lua_scripting_client_cmd(const char *buf)
